% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index-helper.R
\name{check_index2df_order}
\alias{check_index2df_order}
\title{Check input and get urls for index2df(): ordered case}
\usage{
check_index2df_order(board, n, range)
}
\description{
Check input and get urls for index2df(): ordered case
}
\seealso{
Other helper-index2df: \code{\link{check_index2df_search}}
}
\concept{helper-index2df}
\keyword{internal}
