% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index-helper.R
\name{check_index2df_search}
\alias{check_index2df_search}
\title{Check input and get urls for index2df(): search case}
\usage{
check_index2df_search(board, search)
}
\description{
Check input and get urls for index2df(): search case
}
\seealso{
Other helper-index2df: \code{\link{check_index2df_order}}
}
\concept{helper-index2df}
\keyword{internal}
