% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp-df2corp.R
\name{comment2corpus}
\alias{comment2corpus}
\alias{comment2qcorp}
\alias{comment2tmcorp}
\title{Convert 'comment' list-column to 'corpus' list-column}
\usage{
comment2qcorp(df)

comment2tmcorp(df, PCorpus = FALSE, ...)
}
\arguments{
\item{df}{Data frame. A data frame generated by
\code{\link{post2df}}.}

\item{PCorpus}{Logical. Whether to contruct corpus with
\code{\link[tm]{PCorpus}}. If \code{FALSE}, uses
\code{\link[tm]{VCorpus}}. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed on to
\code{\link[quanteda]{corpus}}, \code{\link[tm]{VCorpus}},
or \code{\link[tm]{PCorpus}}.}
}
\description{
These functions convert data frames stored in the list-column
'comment' of the data frame returned by \code{\link{post2df}}
 to corpus objects in \code{\link[quanteda]{corpus}},
 \code{\link[tm]{VCorpus}}, and \code{\link[tm]{PCorpus}}.
}
\seealso{
\code{\link{post2qcorp}}, \code{\link{post2tmcorp}}
}
