% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-retrieve.R
\name{down_html}
\alias{down_html}
\title{Dowload HTML files to local directory}
\usage{
down_html(url, dir, gz = T)
}
\arguments{
\item{url}{Character vector. A list of URLs of posts to
download.}

\item{dir}{Character. Path to the directory to store
dowloaded files.}

\item{gz}{Logical. Whether to compress the downloaded
files. Defaults to \code{TRUE}. Uses gzip compression.}
}
\description{
Dowload HTML files to local directory
}
