% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt-manipulate.R
\name{extr_post_category}
\alias{extr_post_category}
\title{Extract Post Category from Title}
\usage{
extr_post_category(post_title)
}
\arguments{
\item{post_title}{A string with UTF-8 encoding. No leading space
in string.}
}
\value{
If there is square-bracket-match or \code{Re:}-match,
  the function returns a string.
  Else, the function returns string \code{No}.
}
\description{
\code{extr_post_category} returns the category of a PTT post title.
}
\details{
This function extracts the category from the title based on regular
expressions. It extracts the content inside the square brackets
(\code{[, ]}) of the title. If there is no square bracket,
it searches for \code{Re:} and extracts this pattern.
If neither is matched, the function returns
string \code{No}.
}
\examples{

\dontrun{
extr_post_category(str_err)
}

}
\keyword{internal}
