% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_posts.R
\name{get_all_posts}
\alias{get_all_posts}
\title{Extract list of posts from a board}
\usage{
get_all_posts(board_name, max_post = 1000, include.push = FALSE,
  parallel = NULL, ...)
}
\arguments{
\item{board_name}{PTT board name}

\item{max_post}{See get_post_url()}

\item{...}{other parameters passed to get_post_url()}
}
\value{
data.table
}
\description{
Extract list of posts from a board
}
\examples{
get_all_posts("Gossiping", max_post = 10)

}

