% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_post.R
\name{get_all_url}
\alias{get_all_url}
\title{get_url_post}
\usage{
get_all_url(board_name, max_post = 1000L, ...)
}
\arguments{
\item{board_name}{String of PTT board name.}

\item{max_post}{Maximun number of the lastest posts. Default 1000.
If NULL, will get all of listpage_urls. [Caucious!]}
}
\description{
Get urls of posts that listed in the list page.
}
\examples{
post_urls <- get_all_url("Gossiping", max_post = 100)
post_urls

}

