% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{get_board_popular}
\alias{get_board_popular}
\title{Return a data frame with popular boards info}
\source{
\url{https://www.ptt.cc/bbs/hotboards.html}
}
\usage{
get_board_popular()
}
\description{
\code{get_board_popular} returns a data frame of
popular boards (熱門看板) on PTT.
}
\examples{
df <- get_board_popular()
head(df)

}
