% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{get_hotboard_info}
\alias{get_hotboard_info}
\title{Return a data frame with popular boards info}
\source{
\url{https://www.ptt.cc/bbs/hotboards.html}
}
\usage{
get_hotboard_info(get_new = FALSE)
}
\arguments{
\item{get_new}{Logical. Defaults to \code{FALSE}.
If \code{TRUE}, scrapes and retreive data from
\url{https://www.ptt.cc/bbs/hotboards.html}.
If \code{FALSE}, use pre-scraped data stored in
the package.}
}
\description{
\code{get_hotboard_info} returns a data frame of
popular boards (\emph{\enc{熱門看板}{re men kan ban}})
on PTT.
}
\examples{
df <- get_hotboard_info()
head(df)

# Get data update time
attr(df, "date")

}
