% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{get_index_info}
\alias{get_index_info}
\title{Convert a board's index page to data frame}
\usage{
get_index_info(url)
}
\arguments{
\item{url}{A board's index page url.}
}
\description{
\code{get_index_info} takes a board's index url
as input and extract the content of the page into
a data frame with 6 variables.
}
