% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{get_index_url}
\alias{get_index_url}
\title{Get Latest Page url}
\usage{
get_index_url(board_url)
}
\arguments{
\item{board_url}{url.}
}
\value{
Character vector of length 2.
  The first element is a number,
  the second is a url.
}
\description{
Get Latest Page url
}
