% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{get_index_urls}
\alias{get_index_urls}
\title{Get the newest n urls of index page of a board}
\usage{
get_index_urls(board, n)
}
\arguments{
\item{board}{String. Either a url or a board name
that matches one of the entries in \code{board}
of the data frame returned by
\code{\link{get_board_popular}}.}

\item{n}{Number of index page to retreive.}
}
\value{
Chracter vector with length equal to the
  argument \code{n}.
}
\description{
Get the newest n urls of index page of a board
}
