% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-post-helper.R
\name{get_post}
\alias{get_post}
\title{Get all information from an individual PTT post}
\usage{
get_post(post_url, board_col = FALSE)
}
\arguments{
\item{post_url}{Character. An URL of a PTT post.}

\item{board_col}{Logical. Whether to set board name as a
variable. Defaults to \code{FALSE}.}
}
\value{
A data frame with 1 row and 10 variables:
  \describe{
    \item{author}{Author of the post.}
    \item{category}{Category of the post, such as
      "\emph{\href{https://liao961120.github.io/pttR/articles/translation}{xin wen}}",
      "\emph{\href{https://liao961120.github.io/pttR/articles/translation}{wen gua}}",
      "\emph{Re:}".}
    \item{title}{Title of the post.}
    \item{date}{The date of the post.}
    \item{content}{The content of the post.}
    \item{comment}{A list column.
      See \code{\link{get_post_comment}} for information
      about entries in this list column.}
    \item{n_comment}{Number of comments.}
    \item{n_push}{Number of "Push" comments.}
    \item{n_boo}{Number of "Boo" comments.}
    \item{link}{URL of the post with
      \url{https://www.ptt.cc/bbs/} removed.}
  }
  One additional variable is optional:
  \describe{
    \item{board}{The board the post belongs to. Exist only
      if \code{board_col = TRUE}.}
  }
}
\description{
\code{get_post} returns a data frame with 1 row and 9 cols,
where the column \code{comment} is a list column.
}
\keyword{internal}
