% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-post.R
\name{get_post_comment}
\alias{get_post_comment}
\title{Retrieve User Comments from an Individual PTT Post}
\usage{
get_post_comment(post_xml)
}
\arguments{
\item{post_xml}{An \code{xml_document} created by
\code{\link{read_html2}} or \code{\link[xml2]{read_html}}
See \code{\link[xml2]{read_html}} for details.}
}
\value{
A data frame with n rows and 5 variables:
  \describe{
    \item{tag}{tag of the comment, can be one of the 3 values:
      \code{Push} corresponds to "推",
      \code{Boo} corresponds to "噓", and
      \code{Neu} corresponds to "→".}
    \item{user}{ID of the user who left the comment.}
    \item{comment}{The content of the comment.}
    \item{ip}{ip address of the comment.}
    \item{time}{The comment date time.}
  }
}
\description{
\code{get_post_comment} returns a data frame with n rows and 5
cols, where n is the number of comments in the post.
}
\examples{

url <- "https://www.ptt.cc/bbs/Gossiping/M.1534415307.A.BE5.html"
post <- read_html2(url)

get_post_comment(post)

}
