% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-pttweb.R
\name{get_post_content}
\alias{get_post_content}
\title{Retrieve Post Content from an Individual PTT Post}
\usage{
get_post_content(post_xml)
}
\arguments{
\item{post_xml}{An \code{xml_document} created by
\code{\link{read_html2}} or \code{\link[xml2]{read_html}}
See \code{\link[xml2]{read_html}} for details.}
}
\value{
A data frame with 1 row and 1 col. The entry is a long
  string.
}
\description{
\code{get_post_content} returns a data frame with 1 row and 1 col.
}
\examples{

url <- "https://www.ptt.cc/bbs/Gossiping/M.1534415307.A.BE5.html"
post <- read_html2(url)

get_post_content(post)

}
