% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-post.R
\name{get_post_meta}
\alias{get_post_meta}
\title{Retrieve mata data from an individual PTT post}
\usage{
get_post_meta(post_xml, board_col = FALSE)
}
\arguments{
\item{post_xml}{\code{xml_document} created by
\code{\link{read_html2}} or \code{\link[xml2]{read_html}}.
See \code{\link[xml2]{read_html}} for details.}

\item{board_col}{Logical. Whether to set board name as a
variable. Defaults to \code{FALSE}. Note you can get the
board name with \code{attributes{df}$board} or
\code{attr(df, "board")} regardless of the value of this
argument.}
}
\value{
A data frame with 1 row and 4 variables:
  \describe{
    \item{author}{Author of the post.}
    \item{category}{Category of the post, such as
      "\emph{\href{https://liao961120.github.io/pttR/articles/translation}{xin wen}}",
      "\emph{\href{https://liao961120.github.io/pttR/articles/translation}{wen gua}}",
      "\emph{Re:}".}
    \item{title}{Title of the post.}
    \item{date}{The date of the post.}
  }
}
\description{
\code{get_post_meta} returns a data frame with 1 row
and 4 cols.
}
\examples{

#url <- "https://www.ptt.cc/bbs/Gossiping/M.1534415307.A.BE5.html"

#post_meta <- get_post_meta(read_html2(url))
#post_meta

#attributes(post_meta)$board

}
\keyword{internal}
