% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_post.R
\name{get_post_url}
\alias{get_post_url}
\title{get_url_post}
\usage{
get_post_url(listpage_urls, max_post = 1000L, ...)
}
\arguments{
\item{listpage_urls}{The url of the list page.}

\item{max_post}{Maximun number of the lastest posts. Default 1000.
If NULL, will get all of listpage_urls. [Caucious!]}
}
\description{
Get urls of posts that listed in the list page.
}
\examples{
listpage_urls = get_url_listpage("Gossiping")[1:50]
#> [1] "https://www.ptt.cc/bbs/Gossiping/index15002.html"
#> [2] "https://www.ptt.cc/bbs/Gossiping/index15001.html"
#> ...
post_urls = get_post_url(listpage_urls, max_post = 100)
post_urls
}

