% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_listpage.R
\name{get_url_listpage}
\alias{get_url_listpage}
\title{get_url_listpage}
\usage{
get_url_listpage(boardName)
}
\arguments{
\item{boardName}{The name of a board.}
}
\description{
Get urls of the list pages, each contains certain posts of the input board.
}
\examples{
listPageUrls = get_url_listpage("Gossiping")[1:5]
listPageUrls
}

