% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_post.R
\name{get_url_post}
\alias{get_url_post}
\title{get_url_post}
\usage{
get_url_post(listpage_urls)
}
\arguments{
\item{listPageUrl}{The url of the list page.}
}
\description{
Get urls of posts that listed in the list page.
}
\examples{
listpage_urls = get_url_listpage("Gossiping")[1:5]
postUrls = unlist(lapply(listPageUrls,getPostUrls))
postUrls
}

