% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_urls}
\alias{get_urls}
\title{Extract list of urls from a board}
\usage{
get_urls(board_name, max_post = 1000L, mc.cores = -1, ...)
}
\arguments{
\item{board_name}{String of PTT board name.}

\item{max_post}{Maximun number of the lastest posts. Default 1000.
If set to \code{-1}, will get all of listpage_urls. [Caucious!]}
}
\description{
Extract list of urls from a board
}
\examples{
post_urls <- get_urls("Gossiping", max_post = 100)
post_urls

}

