% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{index2df}
\alias{index2df}
\title{Convert a board's index pages to data frame}
\usage{
index2df(board, n = 1, from = NULL, to = NULL)
}
\arguments{
\item{board}{Character. Either a \strong{url} or a
\strong{board name} that matches one of the entries
in the variable \code{board} of the data frame
returned by \code{\link{get_hotboard_info}}, such
as \emph{"Gossiping"}, \emph{"Baseball"}, \emph{"LoL}.
\strong{board name} is case-insensitive. See
\strong{Examples} for details.}

\item{n}{Numeric. Number of pages to scrape.
Defaults to \code{1}, which scrapes only the
newest page. If set to \code{2}, then scrapes
the newest and the second-newest page, and so
forth. \strong{The value should be kept low so
that it doesn't put too much load on the server}.}

\item{from}{Numeric. The starting page of custom range of
index pages to retreive. Must be set together with
\code{to}. When set, argument \code{n} has no effect.
Defaults to \code{NULL}.}

\item{to}{Numeric. The ending page of custom range of
index pages to retreive. See \code{from}.}
}
\value{
Returns a data frame with one post info per
  row.
}
\description{
\code{index2df} scrapes index pages of a board
(\emph{\enc{看板}{kan ban}}) and extracts
information into a data frame.
}
\details{
\code{\link{get_index_info}} extracts data from \emph{one}
index page, while \code{index2df} deals with \emph{several}.
In addition, \code{index2df} returns more information.
Use \code{index2df} instead of \code{\link{get_index_info}}.
}
\examples{
# Get Board Name
head(get_hotboard_info())[, 1]

# Get data from 'Gossiping'
df <- index2df("Gossiping")

\dontrun{
# Or use url directly
link <- "https://www.ptt.cc/bbs/Gossiping/index"

df <- index2df(link, 3)
}

}
\seealso{
\code{\link{get_index_info}}
}
