% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-post.R
\name{mutate_content_len}
\alias{mutate_content_len}
\title{Word count 'content' col of get_post}
\usage{
mutate_content_len(df, idx)
}
\description{
Word count 'content' col of get_post
}
\keyword{internal}
