% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-post.R
\name{mutate_content_url}
\alias{mutate_content_url}
\title{Extract and remove URL from 'content' column of a data frame returned by get_post()}
\usage{
mutate_content_url(df, idx, rpl = "rm_URL")
}
\description{
Extract and remove URL from 'content' column of a data frame returned by get_post()
}
\keyword{internal}
