% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-date.R
\name{parse_comment_date}
\alias{parse_comment_date}
\title{Parse dates to add year in PTT post comments}
\usage{
parse_comment_date(cmt_df, post_date)
}
\arguments{
\item{cmt_df}{A data frame with format almost identical
to that returned by \code{\link{get_post_comment}},
except that the \code{time} column has entries with
the format of the date in PTT post comments.}

\item{post_date}{Date. A value returned by
\code{\link[lubridate]{ymd_hms}}.}
}
\value{
A data frame with dimension same as the input.
}
\description{
To deal with the problem of loss of information,
i.e. a date with its year truncated can't be converted
back to the original date, two assumptions are made:
\enumerate{
  \item The time interval between (1) the time the post
    was published and (2) the time the first comment
    was left is less than 1 year.
  \item The time interval between two adjacent comment
    is less than 1 year.
}
}
\keyword{internal}
