% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-date.R
\name{parse_post_date}
\alias{parse_post_date}
\title{Extract the publish date of a PTT post}
\usage{
parse_post_date(post_xml)
}
\arguments{
\item{post_xml}{\code{xml_document}.}
}
\description{
Extract the publish date of a PTT post
}
\keyword{internal}
