% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp-df2corp.R
\name{post2corpus}
\alias{post2corpus}
\alias{post2qcorp}
\alias{post2tmcorp}
\title{Convert post data frame to corpus objects}
\usage{
post2qcorp(df, ...)

post2tmcorp(df, PCorpus = FALSE, ...)
}
\arguments{
\item{df}{Data frame. A data frame generated by
\code{\link{post2df}}.}

\item{...}{Additional arguments passed on to
\code{\link[quanteda]{corpus}}, \code{\link[tm]{VCorpus}},
or \code{\link[tm]{PCorpus}}.}

\item{PCorpus}{Logical. Whether to contruct corpus with
\code{\link[tm]{PCorpus}}. If \code{FALSE}, uses
\code{\link[tm]{VCorpus}}. Defaults to \code{FALSE}.}
}
\description{
These functions convert data frame to corpus objects
same as those constructed with
\code{\link[quanteda]{corpus}}, \code{\link[tm]{VCorpus}},
and \code{\link[tm]{PCorpus}}. The list-column
\code{comment} of the original post data frame is dropped
for not complicating the data structure of the corpus.
}
\examples{
library(quanteda)
post_df  <- example_posts()
post_corp <- post2qcorp(post_df)

summary(post_corp)

library(tm)
post_df  <- example_posts()
post_corp <- post2tmcorp(post_df)

inspect(post_corp)
meta(post_corp, "author")

}
\seealso{
\code{\link{comment2qcorp}}, \code{\link{comment2tmcorp}}
}
