% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-post.R
\name{post2df}
\alias{post2df}
\title{Extract information from PTT posts}
\usage{
post2df(path, board_col = FALSE)
}
\arguments{
\item{path}{Character Vector. A vector of URLs or local paths
to PTT posts.}

\item{board_col}{Logical. Whether to set board name as a new
variable. Defaults to \code{FALSE}.}
}
\value{
A data frame with n rows and 12 variables:
  \describe{
    \item{author}{Author of the post.}
    \item{category}{Category of the post, such as
      "\emph{\href{https://liao961120.github.io/pttR/articles/translation}{xin wen}}",
      "\emph{\href{https://liao961120.github.io/pttR/articles/translation}{wen gua}}",
      "\emph{Re:}".}
    \item{title}{Title of the post.}
    \item{date}{The date of the post.}
    \item{content}{The content of the post.}
    \item{content_char}{The Number of characters in the post
      content. Whitespaces and newline characters are
      removed before counting.}
    \item{n_comment}{Number of comments.}
    \item{n_push}{Number of "Push" comments.}
    \item{n_boo}{Number of "Boo" comments.}
    \item{link}{URL of the post with
      \url{https://www.ptt.cc/bbs/} removed.
      For local file paths, the link is the file name.}
    \item{comment}{A list-column with data frames stored
      inside. Contents extracted from the post comment
      region. See \code{\link{get_post_comment}} for
      information about the variables in the data frame.}
    \item{comment_urls}{A list-column with character vectors
      stored inside. URLs extracted from post content are stored
      inside the character vectors. The original URLs in the
      post content are replaced as '\emph{rm_URL}' in the
      variable 'content'.}
  }
  One additional variable is optional:
  \describe{
    \item{board}{The board the post belongs to. Exist only
      if \code{board_col = TRUE}.}
   }
}
\description{
\code{post2df} takes paths to PTT posts (either URLs or local
path to HTML files) as input, extracts information from the
posts, and returns a data frame with n row and 12 cols with
one post per row.
}
\details{
This is a function that rbinds the data gathered from
\code{\link{get_post}}, and add some meta data about the
'content' of the post.
}
\examples{
url <- "https://www.ptt.cc/bbs/Gossiping/M.1534415307.A.BE5.html"

post_df <- post2df(url)
head(post_df)

# Access information in the list column: 'comment'
head(post_df$comment[[1]])

\dontrun{
# Read from local files
post_df <- post2df(list.files('local/gossiping', full.names = T))
}

}
