% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.R
\name{seg_ptt}
\alias{seg_ptt}
\title{Word segmentation for PTT post content}
\usage{
seg_ptt(x, words = NULL, tags = NULL, user = NULL)
}
\arguments{
\item{x}{Character vector. Text to be segmented.}
}
\description{
Word segmentation for PTT post content
}
\keyword{internal}
