% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.R
\name{seg_content}
\alias{seg_content}
\alias{seg_comment}
\alias{segment}
\title{Word segmentation for PTT post content and comments.}
\source{
For details about the built-in ptt dictionary,
  see \url{https://liao961120.github.io/PTT-scrapy/}.
}
\usage{
seg_content(x, words = NULL, tags = NULL, user = NULL)

seg_comment(x, words = NULL, tags = NULL, user = NULL)
}
\arguments{
\item{x}{Column 'content' or 'comment' from a data frame
returned by \code{post2df}.}

\item{words}{Character vector. A vector of words to
pass to jiebaR dictionary.
See \code{\link[jiebaR]{new_user_word}} for details.}

\item{tags}{Character vector. A vector of tags
specifying the lexical categories of the words in
`words`. Defaults to `n` (noun).
See \code{\link[jiebaR]{new_user_word}} for details.}

\item{user}{Character. A string specifying the path to
an user defined dictionary. Defaults to pttR built-in
dictionary. See
\href{jiebaR doc}{https://qinwenfeng.com/jiebaR/worker-.html#user-}
for details.}
}
\description{
Word segmentation for PTT post content and comments.
}
