% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt-handy.R
\name{as_url}
\alias{as_url}
\title{Turn PTT board name to URL}
\usage{
as_url(x, pre = "https://www.ptt.cc/bbs/")
}
\arguments{
\item{x}{Character. A board name or a partial URL (ending in
\code{.html}) with base URL removed.}

\item{pre}{Character. A base URL. Defaults to the URL of
\href{https://www.ptt.cc/bbs/}{PTT Web}.}
}
\description{
A wrapper of \code{\link[base]{paste0}} to turn board names
or partial post URLs removed back to the original URLs.
}
\examples{
as_url("gossiping")
as_url("Gossiping/M.1534490816.A.A3A.html")

}
