% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{check_input_board}
\alias{check_input_board}
\title{Input Check for 'board'}
\usage{
check_input_board(board)
}
\description{
Check and get the index url of a board. If invalid,
stops executing and gives error message.
}
\keyword{internal}
