% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_posts.R
\name{get_all_posts}
\alias{get_all_posts}
\title{get_all_posts}
\usage{
get_all_posts(board_name, max_post = 1000, ...)
}
\arguments{
\item{board_name}{PTT board name}

\item{max_post}{See get_post_url()}

\item{...}{other parameters passed to get_post_url()}
}
\value{
data.table
}
\description{
Get all posts from a given board name
}
\examples{
get_all_posts("gissiping")

}

