% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_posts.R
\name{get_all_posts}
\alias{get_all_posts}
\title{get_all_posts}
\usage{
get_all_posts(board_name, ...)
}
\arguments{
\item{board_name}{}
}
\value{
data.table
}
\description{
Get all posts from a given board name
}
\examples{
get_all_posts("gissiping")

}

