% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-post.R
\name{get_post}
\alias{get_post}
\title{Get Information from an Individual PTT Post}
\usage{
get_post(post_xml)
}
\arguments{
\item{post_xml}{An \code{xml_document} created by
\code{\link{read_html2}} or \code{\link[xml2]{read_html}}
See \code{\link[xml2]{read_html}} for details.}
}
\value{
A data frame with 1 row and 9 variables:
  \describe{
    \item{author}{Author of the post.}
    \item{category}{Category of the post, such as
      \emph{\enc{新聞}{xin wen}},
      \emph{\enc{問卦}{wen gua}}, \emph{Re:}.}
    \item{title}{Title of the post.}
    \item{date}{The date of the post.}
    \item{content}{The content of the post.}
    \item{comment}{A list column.
      See \code{\link{get_post_comment}} for information
      about entries in this list column.}
    \item{n_comment}{Number of comments.}
    \item{n_push}{Number of "Push" comments.}
    \item{n_boo}{Number of "Boo" comments.}
  }
}
\description{
\code{get_post} returns a data frame with 1 row and 9 cols,
where the column \code{comment} is a list column.
}
\details{
This is a function that combines the data gathered from
three functions: \code{\link{get_post_meta}},
\code{\link{get_post_content}}, and
\code{\link{get_post_comment}}.
}
\examples{

url <- "https://www.ptt.cc/bbs/Gossiping/M.1534415307.A.BE5.html"
post <- read_html2(url)

post_df <- get_post(post)
head(post_df)

# Access information in the list column: 'comment'
head(post_df$comment[[1]])

}
