% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_content.R
\name{get_post_content}
\alias{get_post_content}
\title{getPostData}
\usage{
get_post_content(post_url, max_error_time = 3)
}
\arguments{
\item{post_url}{URL of a post.}
}
\description{
Get the content of the post.
}
\examples{
get_post_content("https://www.ptt.cc/bbs/Gossiping/M.1467117389.A.62D.html")

}

