% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-post.R
\name{get_post_meta}
\alias{get_post_meta}
\title{Retrieve Mata Data from an Individual PTT Post}
\usage{
get_post_meta(post_xml)
}
\arguments{
\item{post_xml}{\code{xml_document} created by
\code{\link{read_html2}} or \code{\link[xml2]{read_html}}
See \code{\link[xml2]{read_html}} for details.}
}
\value{
A data frame with 1 row and 4 variables:
  \describe{
    \item{author}{Author of the post.}
    \item{category}{Category of the post, such as
      \emph{\enc{新聞}{xin wen}},
      \emph{\enc{問卦}{wen gua}}, \emph{Re:}.}
    \item{title}{Title of the post.}
    \item{date}{The date of the post.}
  }
}
\description{
\code{get_post_meta} returns a data frame with 1 row
and 4 cols.
}
\examples{

url <- "https://www.ptt.cc/bbs/Gossiping/M.1534415307.A.BE5.html"
post <- read_html2(url)

get_post_meta(post)

}
