% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posts_list.R
\name{get_posts_list}
\alias{get_posts_list}
\alias{get_urls}
\alias{get_titles}
\title{Extract list of urls, titles, nrec from a board}
\usage{
get_posts_list(board_name, max_post = 1000L, mc.cores = -1, ...)

get_urls(board_name, max_post = 1000L, mc.cores = -1, ...)

get_titles(board_name, max_post = 1000L, mc.cores = -1, ...)
}
\arguments{
\item{board_name}{String of PTT board name.}

\item{max_post}{Maximun number of the lastest posts. Default 1000.
If set to \code{-1}, will get all of listpage_urls. [Caucious!].}

\item{mc.cores}{Number of parallel cores to use. Use \code{-1} to auto detect.}
}
\description{
Extract list of urls, titles, nrec from a board
}
\section{Functions}{
\itemize{
\item \code{get_urls}: Get urls of posts in a board.

\item \code{get_titles}: Get post titles of posts in a board.
}}

\examples{
get_posts_list("Gossiping", max_post = 100)
get_urls("Gossiping", max_post = 100)
get_titles("Gossiping", max_post = 100)

}
