% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index2df.R
\name{index2df}
\alias{index2df}
\title{Extract data from multiple index pages of a PTT board.}
\usage{
index2df(board, newest = 1, pages = NA, search_term = NA,
  search_page = 1)
}
\arguments{
\item{board}{Character. Either a \strong{url} or a
\strong{board name}, such as \emph{"Gossiping"},
\emph{"Baseball"}, \emph{"LoL"}.
\strong{board name} is case-insensitive. See
\strong{Examples} for details.
\code{board} has a different requirements when used
with argument \code{search} (See below).}

\item{newest}{Integer. Number of pages, starting from
the most recent page, to scrape.
Defaults to \code{1}, which scrapes only the
newest page. If set to \code{2}, then scrapes
the newest and the second-newest page, and so
forth.}

\item{pages}{Integer vector. A vector of index page number(s).
This parameter lets you scrape the index pages
you want, provided that the page exist. Becareful not to
provide numbers exceeding the range of current index pages.
Defaults to \code{NA}.}

\item{search_term}{Character. A term to search in the index,
such as "\emph{魯蛇}". There are also some advanced
search methods:
\describe{
\item{Post thread}{Prepend "\emph{thread:}" to the
search term (post title), e.g.
"\emph{thread:[新聞]抱怨「垃圾不分藍綠」姚文智：害民進黨被}"
.}
\item{Posts of an author}{Prepend "\emph{author:}"
to the author's ID, e.g., "\emph{author:Plumage}".}
}}

\item{search_page}{Integer vector. A vector of index page
number(s). With argument \code{search_term}) set,
\code{search_page} lets you to scrape index pages related
to a specific term. Defaults to \code{1}, which scrapes only
the newest page.}
}
\value{
A data frame with one post info per row.
}
\description{
\code{index2df} scrapes the index pages of a board
(\emph{\enc{看板}{kan ban}}) and extracts the
information into a data frame.
}
\section{Warning}{

\strong{Do not request too many pages one time}.
It places heavy load on the server.
}

\examples{
# Get data from 'Gossiping'
df <- index2df("Gossiping")

\dontrun{
# Or use url directly
link <- "https://www.ptt.cc/bbs/Gossiping/index"

df <- index2df(link)
}
}
\seealso{
\code{\link{get_index_info}}
  \code{\link{get_index_info}} extracts data from
  \emph{one} index page, while \code{index2df} deals with
  \emph{several}. In addition, \code{index2df} has more
  functionality to deal with multiple pages extraction
}
