% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index.R
\name{index2df}
\alias{index2df}
\title{Convert a board's index pages to data frame}
\usage{
index2df(board, n = 1)
}
\arguments{
\item{board}{A string. Either a url or a board name
that matches one of the entries in the variable
\code{board} of the data frame returned by
\code{\link{get_hotboard_info}}. See
\strong{Examples} for details.}

\item{n}{Numeric. Number of pages to scrape.
Defaults to \code{1}, which scrapes only the
newest page. If set to \code{2}, then scrapes
the newest and the second-newest page, and so
forth. \strong{The value should be kept low so
that it doesn't put too much load on the server}.}
}
\value{
Returns a data frame with one post info per
  row.
}
\description{
\code{index2df} scrapes a board's (看板) index page
extracts information into a data frame.
}
\examples{

# Get Board Name
head(get_hotboard_info())[, 1]

# Get data from 'Gossiping'
df <- index2df("Gossiping")

\dontrun{
# Or use url directly
link <- "https://www.ptt.cc/bbs/Gossiping/index"

df <- index2df(link, 3)
}

}
