% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap-pttweb.R
\name{read_html2}
\alias{read_html2}
\title{Read PTT Pages with "over18-confirmation"}
\usage{
read_html2(url)
}
\arguments{
\item{url}{The url of the target web page.}
}
\value{
An XML document. See \code{\link[xml2]{read_html}}
  for more information.
}
\description{
A wrapper of \code{xml2::read_html()} with cookie set
to bypass PTT's "over18-confirmation".
}
\examples{
url <- "https://www.ptt.cc/bbs/Gossiping/index.html"
read_html2(url)
}
