% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-gen-fun.R
\name{read_html2}
\alias{read_html2}
\title{Read PTT pages with "over18-confirmation"}
\usage{
read_html2(url, ...)
}
\arguments{
\item{url}{Character. The URL of the target web page.}

\item{...}{Additional arguments passed on to
\code{\link[xml2]{read_html}}.}
}
\value{
An XML document. See \code{\link[xml2]{read_html}}
  for more information.
}
\description{
A wrapper of \code{xml2::read_html()} with cookie set
to bypass PTT's "over18-confirmation".
}
\examples{
url <- "https://www.ptt.cc/bbs/Gossiping/index.html"
read_html2(url)

}
