% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-index-helper.R
\name{scrape-index}
\alias{scrape-index}
\alias{get_index_url}
\alias{get_index_info}
\title{Helper functions for scraping index pages}
\usage{
get_index_url(board_url)

get_index_info(board_url)
}
\arguments{
\item{board_url}{Character. A board's index page url.}
}
\value{
\code{get_index_url} returns a char vector of
  length 2. The first element is a number,
  and the second is a url.

\code{get_index_info} returns a data frame
  with n rows and 6 variables, where n is the number
  of post links on an index page.
}
\description{
These are helper functions for package development, not
written in man page index but available to advanced
users.

\code{get_index_url} finds out the newest index page
of a board. It takes a board's url (e.g.
\url{https://www.ptt.cc/bbs/Gossiping/index.html})
as input and returns a character vector of length 2.

\code{get_index_info} takes a board's index url
as input and extract the content of the page into
a data frame with 6 variables.
}
\seealso{
\code{\link{extr_post_category}}
}
\keyword{internal}
