% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.R
\name{seg_ptt}
\alias{seg_ptt}
\title{Word segmentation for PTT post content}
\source{
\url{https://liao961120.github.io/PTT-scrapy/}
}
\usage{
seg_ptt(x, words = NULL, tags = NULL, user = NULL)
}
\arguments{
\item{x}{Character vector or data frame. Text to be
segmented. If a data frame is passed, the column
'content' or 'comment' will be extracted and
use as the input character vector.}

\item{words}{Character vector. A vector of words to
pass to jiebaR dictionary.
See \code{\link[jiebaR]{new_user_word}} for details.}

\item{tags}{Character vector. A vector of tags
specifying the lexical categories of the words in
`words`. Defaults to `n` (noun).
See \code{\link[jiebaR]{new_user_word}} for details.}

\item{user}{Character. A string specifying the path to
user defined dictionary. Defaults to pttR built-in
dictionary. See
\href{jiebaR doc}{https://qinwenfeng.com/jiebaR/worker-.html#user-} for details.}
}
\description{
Word segmentation for PTT post content
}
